refactor(awsSns): separate awsSnsCredential from awsCredential
